requirements.txt
acres==0.3.0
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
cohere==5.14.2
configobj==5.0.9
configparser==7.2.0
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
etelemetry==0.3.1
exceptiongroup==1.2.2
fastapi==0.115.12
fastapi-cache2==0.2.2
fastapi-cli==0.0.7
fastavro==1.10.0
filelock==3.18.0
fsspec==2025.3.2
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.1
idna==3.10
importlib_resources==6.5.2
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
looseversion==1.3.0
lxml==5.3.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
nltk==3.9.1
numpy==2.2.4
openai==1.70.0
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pendulum==3.0.0
pillow==11.1.0
pinecone==6.0.2
pinecone-plugin-interface==0.0.7
prov==2.0.1
puremagic==1.28
pydantic==2.11.2
pydantic-extra-types==2.10.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
pydot==3.0.4
Pygments==2.19.1
PyMuPDF==1.25.5
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==6.3.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.1
scipy==1.15.2
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
time-machine==2.16.0
tokenizers==0.21.1
tqdm==4.67.1
traits==7.0.2
typer==0.15.2
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
Jinja2==3.1.6
